+++ tests/testthat/test-openfair.R
@@ -1,7 +1,7 @@
-context("OpenFAIR simulation functions")
+context("Sample TEF")
 test_that("Sample TEF", {
   set.seed(1234)
-  tef <- sample_tef(n = 10, params = list(1, 1, 10, 100))
+  tef <- sample_tef(params = list(10, 1, 10, 100))
   expect_that(tef, is_a("list"))
   # ensure that the list has the required elements
   expect_that(names(tef), equals(c("type", "samples", "details")))
@@ -19,7 +19,7 @@ test_that("Sample TEF", {
 context("Sample DIFF")
 test_that("Sample DIFF", {
   set.seed(1234)
-  dat <- sample_diff(n = 10, params = list(1, 50, 70, 75, 3))
+  dat <- sample_diff(params = list(10, 50, 70, 75, 3))
   expect_that(dat, is_a("list"))
   # ensure that the list has the required elements
   expect_that(names(dat), equals(c("type", "samples", "details")))
@@ -34,9 +34,10 @@ test_that("Sample DIFF", {
   )
 })
 
+context("Sample TC")
 test_that("Sample TC", {
   set.seed(1234)
-  tc <- sample_tef(n = 10, params = list(1, 50, 75, 100, 4))
+  tc <- sample_tef(params = list(10, 50, 75, 100, 4))
   expect_that(tc, is_a("list"))
   # ensure that the list has the required elements
   expect_that(names(tc), equals(c("type", "samples", "details")))
@@ -49,6 +50,37 @@ test_that("Sample TC", {
                                            68.4387021948896, 68.7541469869603,
                                            68.554057026653, 64.9764652390671))
               )
+})
+
+context("Sample VULN")
+test_that("Sample VULN works with binom", {
+  set.seed(1234)
+  dat <- sample_vuln(params = list(10, 1, .5))
+  expect_that(dat, is_a("list"))
+  # ensure that the list has the required elements
+  expect_that(names(dat), equals(c("type", "samples", "details")))
+  # ensure that the samples matches the number requested
+  expect_that(length(dat$samples), equals(10))
+  # ensure that values of samples is correct
+  expect_that(sum(dat$samples), equals(7)
+  )
+})
+test_that("Sample VULN works with TC and DIFF", {
+  set.seed(1234)
+  tc <- sample_tc(params = list(10, 50, 70, 85, 2))$samples
+  diff <- sample_diff(params = list(10, 50, 70, 85, 2))$samples
+  dat <- sample_vuln(func = detail_vuln, params = list(tc = tc, diff = diff))
+  expect_is(dat, "list")
+  # ensure that the list has the required elements
+  expect_equivalent(names(dat), c("type", "samples", "details"))
+  # ensure that the samples matches the number requested
+  expect_equivalent(length(dat$samples), 10)
+  # ensure that values of samples is correct
+  expect_equivalent(sum(dat$samples), 5)
+  # ensure that mean_tc_exceedance is set correctly
+  expect_equivalent(floor(dat$details$mean_tc_exceedance), 7)
+  # ensure that mean_diff_exceedance is set correctly
+  expect_equivalent(floor(dat$details$mean_diff_exceedance), 8)
 })
 
 context("Sample LM")
@@ -75,9 +107,11 @@ test_that("Simulation", {
                             lm_l=1, lm_ml=100, lm_h = 10000, lm_conf=54),
                        diff_estimates = data_frame(l=1, ml=10, h = 50, conf =4),
                        n = 100)
-  expect_that(sim, is_a("tbl_df"))
-  expect_that(nrow(sim), equals(100))
-  expect_that(length(sim), equals(11))
+  expect_s3_class(sim, "tbl_df")
+  expect_equal(nrow(sim), 100)
+  expect_equal(length(sim), 11)
+  expect_equal(sum(sim$threat_events), 2287)
+  expect_equal(sum(sim$loss_events), 781)
 })